% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muskingum.R
\name{calibra_muskingum}
\alias{calibra_muskingum}
\title{Calibra el modelo de Muskingum}
\usage{
calibra_muskingum(i, o, dt, s1)
}
\arguments{
\item{i}{Vector with the inputs}

\item{o}{Vector with the outputs}

\item{dt}{Temporal step}

\item{s1}{Parámetro}
}
\description{
Calibra el modelo de Muskingum
}
