% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaPaleta.R
\name{col2GRASS}
\alias{col2GRASS}
\title{Writes a GRASS colr file}
\usage{
col2GRASS(capa, paleta, intervalos, type = "rast", columna)
}
\arguments{
\item{capa}{GRASS layer}

\item{paleta}{Vector of colors}

\item{intervalos}{Color intervals}

\item{type}{Type of layer ("rast" or "vect"}

\item{columna}{Columna del vectorial que contendrá datos de color}
}
\value{
Paleta de color
}
\description{
Writes a GRASS colr file
}
