% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Función para calcular CM, OA, K, EC y EO}
\usage{
confusionMatrix(pred, ref, labels = NULL)
}
\arguments{
\item{pred}{Predicted values}

\item{ref}{Reference values)}

\item{labels}{Labels for the classes}
}
\value{
The confusion matrix
}
\description{
Función para calcular CM, OA, K, EC y EO
}
