% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaPaleta.R
\name{creaPaleta}
\alias{creaPaleta}
\title{Crea paletas de color}
\usage{
creaPaleta(
  color = "brewer",
  brewer = "Blues",
  n = 5,
  inifin = c("red", "blue")
)
}
\arguments{
\item{color}{Tipo de paleta. Uno de "brewer", "rainbow", "heat", "terrain", "topo", "cm" o "ramp"}

\item{brewer}{Tipo de paleta Brewer}

\item{n}{Número de colores en la paleta}

\item{inifin}{Vector con colores}
}
\value{
Paleta de color
}
\description{
Crea paletas de color
}
