% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_hist.R
\name{cumHist}
\alias{cumHist}
\title{Función para hacer histogramas acumulados}
\usage{
cumHist(
  x,
  probability = FALSE,
  main = NULL,
  xlab = deparse(substitute(x)),
  ylab = NULL
)
}
\arguments{
\item{x}{Vector of values}

\item{probability}{If TRUE shows probability instead of frequency}

\item{main}{Tïtulo principal}

\item{xlab}{Etiqueta para el eje de abcisas}

\item{ylab}{Etiqueta para el eje de ordenadas}
}
\description{
Función para hacer histogramas acumulados
}
