% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{datos}
\alias{datos}
\title{Datos inventados para calcular la moda}
\format{Un simple vector}
\source{
\href{https://www.um.es}{umu}
}
\usage{
data(datos)
}
\description{
Datos inventados para calcular la moda
}
\examples{
data(datos)
moda(datos)


}
\references{
Invento
}
\keyword{datasets}
