% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancia.R, R/spatanal.R
\name{distancia}
\alias{distancia}
\title{Calcula distancia de Minkowski}
\usage{
distancia(x1, x2, y1, y2, p = 2)

distancia(x1, x2, y1, y2, p = 2)
}
\arguments{
\item{x1}{Coordenada X del primer punto}

\item{x2}{Coordenada X del segundo punto}

\item{y1}{Coordenada Y del primer punto}

\item{y2}{Coordenada Y del segundo punto}

\item{p}{Exponente de la distancia de Minkowski}
}
\value{
Minkowski's distance
}
\description{
Calcula distancia de Minkowski

Calcula distancia de Minkowski
}
