% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergencia.R
\name{divergencia}
\alias{divergencia}
\title{Divergencia entre clases (Singh, 1984) ver Tso Mather (2009)}
\usage{
divergencia(mat1, mat2)
}
\arguments{
\item{mat1}{Matriz con las reflectividades de la clase 1 con las bandas en columnas y los píxeles en filas}

\item{mat2}{Matriz con las reflectividades de la clase 2 con las bandas en columnas y los píxeles en filas}
}
\value{
Divergencia
}
\description{
Divergencia entre clases (Singh, 1984) ver Tso Mather (2009)
}
