% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtetmax.R
\name{dSqrtETmax}
\alias{dSqrtETmax}
\title{Función de densidad sqrt-ET max}
\usage{
dSqrtETmax(x, k, a)
}
\arguments{
\item{x}{A vector of x values}

\item{k}{Parameter k}

\item{a}{Parameter a}
}
\value{
mass densitiy function
}
\description{
Función de densidad sqrt-ET max
}
