% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejecuta.R
\name{ejecuta}
\alias{ejecuta}
\title{Executes an Operating System Command}
\usage{
ejecuta(command, intern = FALSE)
}
\arguments{
\item{command}{Command to be executed}

\item{intern}{If TRUE the outpus of the commasnd is passed into R.}
}
\value{
If intern=TRUE returns the oputput of the OS command, else returns the command.
}
\description{
Executes an Operating System Command
}
