% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.R
\name{EMC_ahp}
\alias{EMC_ahp}
\title{Proceso de Jerarquía Analítica}
\usage{
EMC_ahp(matriz)
}
\arguments{
\item{matriz}{Matriz de preferencias}
}
\value{
Un vector con los pesos y la razón de consistencia.
}
\description{
Proceso de Jerarquía Analítica
}
