% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.R
\name{EMC_combinacion}
\alias{EMC_combinacion}
\title{Calculo de combinaci贸n lineal en EMC.}
\usage{
EMC_combinacion(matriz, w = NULL)
}
\arguments{
\item{matriz}{Matriz que contiene en filas las alternativas y en columnas los criterios}

\item{w}{Vector con los coeficientes de ponderaci贸n de los criterios. Por defecto ponderan todos igual.}
}
\value{
Un vector con la adecuaci贸n de cada alternativa.
}
\description{
Calculo de combinaci贸n lineal en EMC.
}
