% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.R
\name{EMC_pesos_criterios}
\alias{EMC_pesos_criterios}
\title{Calculo de pesos para diferentes criterios. Si solo se le pasan los criterios utilizaa jerarquía analítica, si se le pasan las importancia utiliza tasación simple y si se le pasa la matriz utiliza el proceso de jerarquía analítica.}
\usage{
EMC_pesos_criterios(criterios, importancia = NULL, matriz = NULL)
}
\arguments{
\item{criterios}{Vector con los nombres de los criterios en orden de importancia}

\item{importancia}{Vector con los valores de importancia absoluta de cada criterio}

\item{matriz}{Matriz de comparación por pares}
}
\value{
Un vector con los pesos salvo cuando se utiliza AHP, en este caso se devuelve también la razón de consistencia.
}
\description{
Calculo de pesos para diferentes criterios. Si solo se le pasan los criterios utilizaa jerarquía analítica, si se le pasan las importancia utiliza tasación simple y si se le pasa la matriz utiliza el proceso de jerarquía analítica.
}
