% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.R
\name{EMC_topsis}
\alias{EMC_topsis}
\title{Calculo de distancia al punto ideal, antiideal y método Topsis.}
\usage{
EMC_topsis(matriz, pi = NULL, pai = NULL, w = NULL)
}
\arguments{
\item{matriz}{Matriz que contiene en filas las alternativas y en columnas los criterios}

\item{pi}{Vector con los valores del punto ideal. Por defecto se toma el máximo de cada criterio}

\item{pai}{Vector con los valores del punto antiideal. Por defecto se toma el mínimo de cada criterio}

\item{w}{Vector con los coeficientes de ponderación de los criterios. Por defecto ponderan todos igual.}
}
\value{
Un data.frame con la distancia a los puntos ideal y antiideal y el resultados de Topsis.
}
\description{
Calculo de distancia al punto ideal, antiideal y método Topsis.
}
