% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.R
\name{EMC_var2crit}
\alias{EMC_var2crit}
\title{Convierte variables en criterios}
\usage{
EMC_var2crit(x, var, cri = NULL, from = NULL, to = NULL)
}
\arguments{
\item{x}{Vector con la variable a transformar}

\item{var}{Vector con los valores de la variable en los puntos de inflexi칩n del gr치fico}

\item{cri}{Vector con los valores del criterio en los puntos de inflexi칩n del gr치fico}

\item{from}{Vector con los valores de la variable para reclasificar}

\item{to}{Vector con los valores del criterio tras reclasificar}
}
\value{
Vector con los valores de criterio
}
\description{
Convierte variables en criterios
}
