% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_thornthwaite.R
\name{et_Thornthwaite}
\alias{et_Thornthwaite}
\title{ET0  Thornthwaite mensuales}
\usage{
et_Thornthwaite(temp, hsm)
}
\arguments{
\item{temp}{Vector con 12 temperaturas medias mensuales}

\item{hsm}{Vector con 12 horas de sol medias mensuales}
}
\value{
Vector con 12 valores de ET0 mensuales
}
\description{
ET0  Thornthwaite mensuales
}
