% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fun.CVLM}
\alias{fun.CVLM}
\title{Función para validación cruzada uno a uno de regresión lineal}
\usage{
fun.CVLM(variable, predictores, data)
}
\arguments{
\item{variable}{Variable dependiente}

\item{predictores}{Vector con los predictores que vamos a utilizar}

\item{data}{Data.frame con los datos}
}
\description{
Función para validación cruzada uno a uno de regresión lineal
}
