% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fun.CVRK}
\alias{fun.CVRK}
\title{Funci贸n para validaci贸n cruzada uno a uno de Regression(lm)-krigging}
\usage{
fun.CVRK(variable, predictores, data)
}
\arguments{
\item{variable}{Variable dependiente}

\item{predictores}{Vector con los predictores que vamos a utilizar}

\item{data}{Data.frame con los datos}
}
\description{
Funci贸n para validaci贸n cruzada uno a uno de Regression(lm)-krigging
}
