% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidro_mod.R
\name{genera_lluvia}
\alias{genera_lluvia}
\title{Genera series de precipitación diaria a partir de un modelo markoviano y una distribución exponencial}
\usage{
genera_lluvia(m01, m11, pexp, n = 365)
}
\arguments{
\item{m01}{Vector de probabilidad de día de lluvia tras día seco mensual}

\item{m11}{Vector de probabilidad de día lluvioso tras día lluvioso mensual}

\item{pexp}{Vector de párametro exponencial mensual}

\item{n}{Number of days

export}
}
\description{
Genera series de precipitación diaria a partir de un modelo markoviano y una distribución exponencial
}
