% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofs.R
\name{gofs}
\alias{gofs}
\title{Function to calculate goodness of fit statistics}
\usage{
gofs(obs, pred)
}
\arguments{
\item{obs}{Vector with observed data}

\item{pred}{Vector with predicted data}
}
\value{
List with the values of rmse, r, nse and pbias
}
\description{
Function to calculate goodness of fit statistics
}
\keyword{statistics}
