% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficoPrecEf.R
\name{graficoPrecEf}
\alias{graficoPrecEf}
\title{Crea un gráfico con precipitación, interceptación, infiltración y precipitación efectiva.}
\usage{
graficoPrecEf(
  p,
  interc,
  infil,
  esco,
  q = NULL,
  legpos = "topright",
  col = NULL,
  leg = NULL,
  xlim = NULL,
  qq = NULL
)
}
\arguments{
\item{p}{Precipitación}

\item{interc}{Interceptación}

\item{infil}{Infiltración}

\item{esco}{Escorrentía}

\item{q}{No sé}

\item{legpos}{Posición de la leyenda}

\item{col}{Paleta}

\item{leg}{Leyenda}

\item{xlim}{Límite X}

\item{qq}{No sé

export}
}
\description{
Crea un gráfico con precipitación, interceptación, infiltración y precipitación efectiva.
}
