% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidro_sep.R
\name{hidroSep}
\alias{hidroSep}
\title{Separación de escorrentía base y escorrentía directa en un hidrograma}
\usage{
hidroSep(
  Q,
  t = NULL,
  t1 = NULL,
  t2 = NULL,
  metodo = NULL,
  xlab = "Tiempo",
  ylab = "Q",
  plot = TRUE,
  q0 = NULL,
  t3 = NULL,
  t4 = NULL,
  q3 = NULL,
  q4 = NULL
)
}
\arguments{
\item{Q}{Vector de caudales}

\item{t}{Vector de tiempos}

\item{t1}{Tiempo 1}

\item{t2}{Tiempo 2}

\item{metodo}{Método de separación}

\item{xlab}{Rótulo para eje X}

\item{ylab}{Rótulo para eje Y}

\item{plot}{TRUE/FALSE}

\item{q0}{Caudal 0 para el método 5}

\item{t3}{Tiempo 3 para el método 5}

\item{t4}{Tiempo 4 para el método 5}

\item{q3}{Caudal 3 para el método 5}

\item{q4}{Caudal 4 para el método 5}
}
\description{
Separación de escorrentía base y escorrentía directa en un hidrograma
}
