% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaPaleta.R
\name{HistPal}
\alias{HistPal}
\title{Coloured Histogram}
\usage{
HistPal(
  z,
  colores,
  intervals,
  main = "",
  breaks = "Sturges",
  thres = FALSE,
  xlab = ""
)
}
\arguments{
\item{z}{Data}

\item{colores}{Palette}

\item{intervals}{Object with intervals}

\item{main}{Main text}

\item{breaks}{System to calculate histogram breaks}

\item{thres}{TRUE/FALSE}

\item{xlab}{Xlab text}
}
\description{
Coloured Histogram
}
