% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_conv.R
\name{hu_conv}
\alias{hu_conv}
\title{Convoluci칩n del hietograma con un hidrograma unitariopara obtener el hidrograma}
\usage{
hu_conv(p, hu, area, dt)
}
\arguments{
\item{p}{es el hietograma en mm}

\item{hu}{es el HU como proporci칩n}

\item{area}{es el 치rea de la cuenca en km2}

\item{dt}{es el paso temporal de p y hu en horas}
}
\value{
hidrograma de crecida en m3/sec
}
\description{
Convoluci칩n del hietograma con un hidrograma unitariopara obtener el hidrograma
}
