% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_conv.R
\name{HUdeconv}
\alias{HUdeconv}
\title{Deconvolución del hidrograma para obtener un HU}
\usage{
HUdeconv(p, q, prop = FALSE)
}
\arguments{
\item{p}{es el hietograma en mm}

\item{q}{es el hidrograma en m3/s}

\item{prop}{Si prop==TRUE devuelve el HU como proporciones, si prop==FALSE devuelve el HU como m3/s/mm}
}
\value{
hidrograma unitario instantaneo
}
\description{
Deconvolución del hidrograma para obtener un HU
}
