% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{id}
\alias{id}
\title{Calcula la intensidad máxima en mm/h para una duración de dt en un episodio con intensidad máxima en 24 horas de I24 y con un valor de I1/I24 igual a I1I24.}
\usage{
id(I24, I1I24, dt = 1)
}
\arguments{
\item{I24}{Intensidad máxima en 24 horas}

\item{I1I24}{Valor de I1/I24}

\item{dt}{Duración para la que se calcula la intensidad máxima}
}
\value{
An intensity value
}
\description{
Calcula la intensidad máxima en mm/h para una duración de dt en un episodio con intensidad máxima en 24 horas de I24 y con un valor de I1/I24 igual a I1I24.
}
