% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{IDF}
\alias{IDF}
\title{Obtiene un hietograma de diseño para un episodio con intensidad máxima en 24 horas de I24, un valor de I1/I24 igual a I1I24 y una duración de D horas. La resolución temporal es de dt horas. El parámetro normal indica si el hietograma resultante debe normalizarse para que su suma sea igual a I24.}
\usage{
IDF(I24, I1I24, dt = 1, D = 24, normal = TRUE)
}
\arguments{
\item{I24}{Intensidad máxima en 24 horas}

\item{I1I24}{Valor de I1/I24}

\item{dt}{Incremento temporal}

\item{D}{duración en horas}

\item{normal}{Indica si el hietograma resultante debe normalizarse para que su suma sea igual a I24.}
}
\value{
Un hietograma
}
\description{
Obtiene un hietograma de diseño para un episodio con intensidad máxima en 24 horas de I24, un valor de I1/I24 igual a I1I24 y una duración de D horas. La resolución temporal es de dt horas. El parámetro normal indica si el hietograma resultante debe normalizarse para que su suma sea igual a I24.
}
