% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_tin.R
\name{interp_tin}
\alias{interp_tin}
\title{Interpolate the z value inside a TIN defined by 3 x,y,z points. If the point is not in the triangle returns NA}
\usage{
interp_tin(x, y, z, x0, y0)
}
\arguments{
\item{x}{vector with 3 x values}

\item{y}{vector with 3 y values}

\item{z}{vector with 3 z values}

\item{x0}{Coordinate of the point to interpolate}

\item{y0}{Coordinate of the point to interpolate}
}
\value{
z for the point (x,y) or NA if the point is outside
}
\description{
Interpolate the z value inside a TIN defined by 3 x,y,z points. If the point is not in the triangle returns NA
}
