% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalos.R
\name{intervalos}
\alias{intervalos}
\title{Función para calcular intervalos de clase para una capa raster}
\usage{
intervalos(
  capa,
  n,
  style = "quantile",
  grass = FALSE,
  column = 1,
  fixed = NULL
)
}
\arguments{
\item{capa}{Layer's name}

\item{n}{Number of intervals}

\item{style}{One of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks" or "dpih"}

\item{grass}{If TRUE the layer is loaded from the GRASS location}

\item{column}{Column of the object that contains the data}

\item{fixed}{Fixed intervals in case style="fixed"}
}
\value{
Colour intervals
}
\description{
Función para calcular intervalos de clase para una capa raster
}
