% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm.R
\name{jm}
\alias{jm}
\title{Distancia de Jeffries-Matusita ver tesis Fulggen}
\usage{
jm(mat1, mat2)
}
\arguments{
\item{mat1}{Matriz con las reflectividades de la clase 1 con las bandas en columnas y los píxeles en filas}

\item{mat2}{Matriz con las reflectividades de la clase 2 con las bandas en columnas y los píxeles en filas}
}
\value{
Distancia
}
\description{
Distancia de Jeffries-Matusita ver tesis Fulggen
}
