% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapica.R
\name{largerBbox}
\alias{largerBbox}
\title{Funcion para modificar un bounding box}
\usage{
largerBbox(bb, l)
}
\arguments{
\item{bb}{bounding box a modificar}

\item{l}{margen a anadir (en metros)}
}
\description{
Funcion para modificar un bounding box
}
