% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapica.R
\name{mapica}
\alias{mapica}
\title{Funcion para hacer mapas}
\usage{
mapica(
  pdfFile,
  other = NULL,
  bbox = NULL,
  scale = NULL,
  paper = NULL,
  mai = NULL,
  cex.axis = 1,
  cex.lab = 1,
  cex.leg = 1,
  xlab = "X",
  ylab = "Y",
  type = "base",
  SGDF = NULL,
  raster = NULL,
  paleta = NULL,
  n = 5,
  mx = NULL,
  my = NULL,
  show = FALSE,
  what = "both",
  legtext = NULL,
  breaks = NULL,
  col = 1,
  portrait = FALSE,
  inifin = NULL,
  grid = NULL,
  legend = NULL,
  bg.leg = "white",
  style = "kmeans"
)
}
\arguments{
\item{pdfFile}{Output pdf file}

\item{other}{Other plot instructions}

\item{bbox}{Alternative bbox}

\item{scale}{Map scale}

\item{paper}{c(ancho, alto) en cm o iso}

\item{mai}{Fix margins}

\item{cex.axis}{Axis annotation size}

\item{cex.lab}{Axis label size}

\item{cex.leg}{Legend labels size}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{type}{"base" or "SGDF"}

\item{SGDF}{Name of the raser map}

\item{raster}{raster}

\item{paleta}{Colour palette}

\item{n}{Number of intervals in color palette}

\item{mx}{Separacion en X desde la esquina inferior izquierda (cm)}

\item{my}{Separacion en Y desde la esquina inferior izquierda (cm)}

\item{show}{If TRUE show the PDF}

\item{what}{what is to plot}

\item{legtext}{Legend labels}

\item{breaks}{palette breaks}

\item{col}{Column to represent}

\item{portrait}{Portrait option}

\item{inifin}{Initial and final colors for palette}

\item{grid}{Grid definition}

\item{legend}{Legend position (word or two coordinates vector)}

\item{bg.leg}{Legend backgroud color}

\item{style}{Class interval style}
}
\description{
Funcion para hacer mapas
}
