% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoriaR.R
\name{memoriaR}
\alias{memoriaR}
\title{Lista los objetos en memoria junto a su tamaño por orden de tamaño}
\usage{
memoriaR(format = "b")
}
\arguments{
\item{format}{"b", "Kb", "Mb", "Gb"

return Una lista de los objetos en memoria junto a su tamaño por orden de tamaño}
}
\description{
Lista los objetos en memoria junto a su tamaño por orden de tamaño
}
