% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{Function to calculate mode}
\usage{
moda(v, na.rm = FALSE)
}
\arguments{
\item{v}{Vector with the data}

\item{na.rm}{If TRUE NAs are removed before calculations}
}
\value{
Numeric value with the mode of v
}
\description{
Function to calculate mode
}
\examples{
moda(c(2,3,5,3,2,6,2,3,5,3))

}
\keyword{statistics}
