% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{mode}
\alias{mode}
\title{Function to calculate mode}
\usage{
mode(v)
}
\arguments{
\item{v}{Vector with the data}
}
\value{
Numeric value with the mode of v
}
\description{
Function to calculate mode
}
\examples{
mode(c(2,3,5,3,2,6,2,3,5,3))

}
\keyword{statistics}
