% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muskingum.R
\name{muskingum}
\alias{muskingum}
\title{Calibra el modelo de Muskingum}
\usage{
muskingum(i, k, x, o1, dt)
}
\arguments{
\item{i}{Vector with the inputs}

\item{k}{Parameter k}

\item{x}{Parameter x}

\item{o1}{Parámetro}

\item{dt}{Temporal step}
}
\description{
Calibra el modelo de Muskingum
}
