% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatanal.R
\name{NearestNeig}
\alias{NearestNeig}
\title{Funcion de vecino más próximo.}
\usage{
NearestNeig(X, A = NULL, d = NULL)
}
\arguments{
\item{X}{una matriz con dos columnas (x e y)}

\item{A}{es el área. Si no se incluye se calcula como      A=(max(x)-min(x))*(max(y)-min(y))}

\item{d}{Distancia}
}
\value{
Una lista con la distancia al vecino más próximo de cada punto, d1m,d1a y R1. Si se le pasan unas distancias de  referencia, devuelve los valores de F(d)
}
\description{
Funcion de vecino más próximo.
}
