% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaPaleta.R
\name{pintaPal}
\alias{pintaPal}
\title{Draw a palette}
\usage{
pintaPal(col, border = "black", intervals = NULL, main = "", horiz = TRUE)
}
\arguments{
\item{col}{Paleta}

\item{border}{Color del borde}

\item{intervals}{Intevalos de la paleta}

\item{main}{Título}

\item{horiz}{TRUE/FALSE}
}
\description{
Draw a palette
}
