% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_fit.R
\name{plot_dist_fit}
\alias{plot_dist_fit}
\title{Gráfico de ajuste a una función de distribución}
\usage{
plot_dist_fit(x, dist)
}
\arguments{
\item{x}{Datos}

\item{dist}{Función a probar ("pnorm","plnorm","ppearsonIII","pgumbel")}
}
\value{
Gráfico de ajuste de unos datos a una función de distribucion
}
\description{
Gráfico de ajuste a una función de distribución
}
