% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatanal.R
\name{PointAnal}
\alias{PointAnal}
\title{Basic spatial statistics}
\usage{
PointAnal(x, y)
}
\arguments{
\item{x}{Vector of X coordinates}

\item{y}{Vector of Y coordinates}
}
\value{
x mean, y mean and sds
}
\description{
Basic spatial statistics
}
