% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtetmax.R
\name{pSqrtETmax}
\alias{pSqrtETmax}
\title{Funci贸n de distribuci贸n sqrt-ET max}
\usage{
pSqrtETmax(x, k, a)
}
\arguments{
\item{x}{A vector of x values}

\item{k}{Parameter k}

\item{a}{Parameter a}
}
\value{
distribution function
}
\description{
Funci贸n de distribuci贸n sqrt-ET max
}
