% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtetmax.R
\name{qSqrtETmax}
\alias{qSqrtETmax}
\title{Funci贸n de cuantiles de la distribuci贸n sqrt-ET max}
\usage{
qSqrtETmax(q, k, a)
}
\arguments{
\item{q}{Cuantile}

\item{k}{Parameter k}

\item{a}{Parameter a}
}
\value{
x value corresponding to the q quantile
}
\description{
Funci贸n de cuantiles de la distribuci贸n sqrt-ET max
}
