% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_scale.R
\name{realScale}
\alias{realScale}
\title{Cálculo de escala}
\usage{
realScale(scale, size = "A4", xlim, ylim)
}
\arguments{
\item{scale}{Scale's denominator}

\item{size}{Paper size (mm or ISO)}

\item{xlim}{X limit}

\item{ylim}{Y limit}
}
\description{
Cálculo de escala
}
