% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_margin.R
\name{rf_margin}
\alias{rf_margin}
\title{Función para calcular el margen de Random Forest}
\usage{
rf_margin(predVotos)
}
\arguments{
\item{predVotos}{Votes for each class}
}
\value{
Difference between the two most voted classes
}
\description{
Función para calcular el margen de Random Forest
}
