% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oif.R
\name{RS_oif}
\alias{RS_oif}
\title{Calcula el factor del índice óptimo para una imagen de satélite en un objeto raster}
\usage{
RS_oif(rast)
}
\arguments{
\item{rast}{Objeto raster}
}
\value{
Data.frame with band names and OIF
}
\description{
Calcula el factor del índice óptimo para una imagen de satélite en un objeto raster
}
