% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oif.R
\name{RS_raster_kmeans}
\alias{RS_raster_kmeans}
\title{Aplica k-means a un objeto raster y le añade una capa con la clasificación resultante}
\usage{
RS_raster_kmeans(rast, centers, sample = 10000, iter.max = 100)
}
\arguments{
\item{rast}{Objeto raster}

\item{centers}{Número de clases}

\item{sample}{Número de celdillas a incluir en el muestreo}

\item{iter.max}{Número máximo de iteraciones}
}
\value{
The input raster object with a new layer that includes the classification.
}
\description{
Aplica k-means a un objeto raster y le añade una capa con la clasificación resultante
}
