% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oif.R
\name{RS_rescale}
\alias{RS_rescale}
\title{Rescale the layers of a raster object}
\usage{
RS_rescale(rast, to)
}
\arguments{
\item{rast}{Objeto raster}

\item{to}{Vector with the range of the resulting layers}
}
\value{
A new raster object with the layers rescaled
}
\description{
Rescale the layers of a raster object
}
