% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtetmax.R
\name{rSqrtETmax}
\alias{rSqrtETmax}
\title{Función de simulación de la distribución sqrt-ET max}
\usage{
rSqrtETmax(n, k, a)
}
\arguments{
\item{n}{Number of results}

\item{k}{Parameter k}

\item{a}{Parameter a}
}
\value{
n numbers from an sqrt-ET max distribution
}
\description{
Función de simulación de la distribución sqrt-ET max
}
