% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoestadistica.R
\name{semivariograma}
\alias{semivariograma}
\title{Cálcula semivariogramas teóricos}
\usage{
semivariograma(tipo, c0, c1, a, h, amp, freq = 1)
}
\arguments{
\item{tipo}{Tipo de semivariograma}

\item{c0}{Pepita}

\item{c1}{Meseta}

\item{a}{rango}

\item{h}{distancias

return Semivariograma teórico

export}

\item{amp}{Amplitud para el semivariograma de efecto agujero}

\item{freq}{Algo del efecto agujero}
}
\description{
Cálcula semivariogramas teóricos
}
