% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeObjects.R
\name{sizeObjects}
\alias{sizeObjects}
\title{List of objects and its size:}
\usage{
sizeObjects(obs)
}
\arguments{
\item{obs}{The return of ls()}
}
\value{
Paleta de color
}
\description{
List of objects and its size:
}
