% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtetmax.R, R/sqrtetmaxparam.R
\name{SqrtETmaxPar}
\alias{SqrtETmaxPar}
\title{Evaluación de parámetros de la distribución sqrt-ET max}
\usage{
SqrtETmaxPar(m, cv)

SqrtETmaxPar(m, cv)
}
\arguments{
\item{m}{Media}

\item{cv}{Coeficiente de variación

return sqrt-ET max distribution parameters}
}
\value{
sqrt-ET max distribution parameters
}
\description{
Evaluación de parámetros de la distribución sqrt-ET max

Evaluación de parámetros de la distribución sqrt-ET max
}
