% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidro_mod.R
\name{Temez}
\alias{Temez}
\title{Modelo de Témez}
\usage{
Temez(P, ETP, Hmax, Imax, C, Hprev)
}
\arguments{
\item{P}{Precipitación}

\item{ETP}{ETP}

\item{Hmax}{Hmax}

\item{Imax}{Imax}

\item{C}{Parámetro C}

\item{Hprev}{Humedad previa

export}
}
\description{
Modelo de Témez
}
