% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGDAl_selCol.R
\name{writeGDAL_selCol}
\alias{writeGDAL_selCol}
\title{Escribe una columna de un SpatialGridDataFrame como una capa raster}
\usage{
writeGDAL_selCol(SGDF, file, col)
}
\arguments{
\item{SGDF}{El SpatialGridDataFrame}

\item{file}{Fichero que contendrá los datos}

\item{col}{Columna que contiene la capa que se guarda. Puede pasarse su nombre o su número de orden}
}
\description{
Escribe una columna de un SpatialGridDataFrame como una capa raster
}
